abstracted edit datapack to its own file
